independent mims module.
